+++ sgl-model-gateway/src/middleware.rs
@@ -15,6 +15,8 @@ use axum::{
     middleware::Next,
     response::{IntoResponse, Response},
 };
+use bytes::Bytes;
+use http_body::Frame;
 use rand::Rng;
 use subtle::ConstantTimeEq;
 use tokio::sync::{mpsc, oneshot};
@@ -30,7 +32,6 @@ use crate::{
     },
     routers::error::extract_error_code_from_response,
     server::AppState,
-    utils::http_utils::AttachedBody,
     wasm::{
         module::{MiddlewareAttachPoint, WasmModuleAttachPoint},
         spec::{
@@ -43,31 +44,33 @@ use crate::{
     },
 };
 
-/// A guard that holds a token and returns it when dropped.
+/// A body wrapper that holds a token and returns it when the body is fully consumed or dropped.
 /// This ensures that for streaming responses, the token is only returned after the entire
 /// stream has been sent to the client.
-pub struct TokenGuard {
+pub struct TokenGuardBody {
+    inner: Body,
     /// The token bucket to return tokens to. Uses Option so we can take() on drop.
     token_bucket: Option<Arc<TokenBucket>>,
     /// Number of tokens to return.
     tokens: f64,
 }
 
-impl TokenGuard {
-    /// Create a new TokenGuard that will return tokens when dropped.
-    pub fn new(token_bucket: Arc<TokenBucket>, tokens: f64) -> Self {
+impl TokenGuardBody {
+    /// Create a new TokenGuardBody that will return tokens when dropped.
+    pub fn new(inner: Body, token_bucket: Arc<TokenBucket>, tokens: f64) -> Self {
         Self {
+            inner,
             token_bucket: Some(token_bucket),
             tokens,
         }
     }
 }
 
-impl Drop for TokenGuard {
+impl Drop for TokenGuardBody {
     fn drop(&mut self) {
         if let Some(bucket) = self.token_bucket.take() {
             debug!(
-                "TokenGuard: stream ended, returning {} tokens to bucket",
+                "TokenGuardBody: stream ended, returning {} tokens to bucket",
                 self.tokens
             );
             // Use lock-free sync return - no runtime needed, guaranteed token return
@@ -76,7 +79,26 @@ impl Drop for TokenGuard {
     }
 }
 
-pub type TokenGuardBody = AttachedBody<TokenGuard>;
+impl http_body::Body for TokenGuardBody {
+    type Data = Bytes;
+    type Error = axum::Error;
+
+    fn poll_frame(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+    ) -> Poll<Option<Result<Frame<Self::Data>, Self::Error>>> {
+        let this = self.get_mut();
+        Pin::new(&mut this.inner).poll_frame(cx)
+    }
+
+    fn is_end_stream(&self) -> bool {
+        self.inner.is_end_stream()
+    }
+
+    fn size_hint(&self) -> http_body::SizeHint {
+        self.inner.size_hint()
+    }
+}
 
 #[derive(Clone)]
 pub struct AuthConfig {
@@ -493,7 +515,7 @@ pub async fn concurrency_limit_middleware(
         // This ensures that for streaming responses, the token is only returned
         // after the entire stream has been sent to the client.
         let (parts, body) = response.into_parts();
-        let guarded_body = TokenGuardBody::new(body, TokenGuard::new(token_bucket, 1.0));
+        let guarded_body = TokenGuardBody::new(body, token_bucket, 1.0);
         Response::from_parts(parts, Body::new(guarded_body))
     } else {
         // No tokens available, try to queue if enabled
@@ -531,7 +553,7 @@ pub async fn concurrency_limit_middleware(
                             // Wrap the response body with TokenGuardBody to return token when stream ends
                             let (parts, body) = response.into_parts();
                             let guarded_body =
-                                TokenGuardBody::new(body, TokenGuard::new(token_bucket, 1.0));
+                                TokenGuardBody::new(body, token_bucket, 1.0);
                             Response::from_parts(parts, Body::new(guarded_body))
                         }
                         Ok(Err(status)) => {
