+++ sgl-model-gateway/src/routers/grpc/regular/stages/generate/response_processing.rs
@@ -15,7 +15,7 @@ use crate::{
             regular::{processor, streaming},
         },
     },
-    utils::http_utils::AttachedBody,
+    core::AttachedBody,
 };
 
 /// Generate response processing stage
