--- sgl-model-gateway/src/utils/mod.rs
@@ -1,1 +1,0 @@
-pub mod http_utils;
